+++ src/ralph_assets/models_assets.py
@@ -16,6 +16,7 @@ from dateutil.relativedelta import relativedelta
 
 from dj.choices import Country
 from django.contrib.auth.models import User
+from django.core.exceptions import ValidationError
 from lck.django.choices import Choices
 from lck.django.common import nested_commit_on_success
 from lck.django.common.models import (
@@ -67,6 +68,12 @@ ASSET_HOSTNAME_TEMPLATE = getattr(settings, 'ASSET_HOSTNAME_TEMPLATE', None)
 if not ASSET_HOSTNAME_TEMPLATE:
     raise ImproperlyConfigured('"ASSET_HOSTNAME_TEMPLATE" must be specified.')
 HOSTNAME_FIELD_HELP_TIP = getattr(settings, 'HOSTNAME_FIELD_HELP_TIP', '')
+
+INVALID_SERVER_ROOM = 1
+INVALID_RACK = 2
+INVALID_ORIENTATION = 3
+INVALID_POSITION = 4
+REQUIRED_SLOT_NUMBER = 5
 
 
 def _replace_empty_with_none(obj, fields):
@@ -135,6 +142,20 @@ class Orientation(Choices):
     WIDTH = Choices.Group(100)
     left = _("left")
     right = _("right")
+
+    @classmethod
+    def is_width(cls, orientation):
+        is_width = orientation in set(
+            [getattr(cls, key) for key in ['left', 'right']]
+        )
+        return is_width
+
+    @classmethod
+    def is_depth(cls, orientation):
+        is_depth = orientation in set(
+            [getattr(cls, key) for key in ['front', 'back', 'middle']]
+        )
+        return is_depth
 
 
 class AssetType(Choices):
@@ -918,16 +939,17 @@ class Rack(Named.NonUnique):
         unique_together = ('name', 'data_center')
 
     data_center = models.ForeignKey(DataCenter, null=False, blank=False)
-    deprecated_ralph_rack = models.ForeignKey(
-        DeprecatedRalphRack, null=True, related_name='deprecated_asset_rack',
-        blank=True,
-    )
     max_u_height = models.IntegerField(default=48)
     server_room = models.ForeignKey(
         ServerRoom, verbose_name=_("server room"),
         null=True,
         blank=True,
     )
+    deprecated_ralph_rack = models.ForeignKey(
+        DeprecatedRalphRack, null=True, related_name='deprecated_asset_rack',
+        blank=True,
+    )
+
 
 
 class DeviceInfo(TimeTrackable, SavingUser, SoftDeletable):
@@ -941,8 +963,8 @@ class DeviceInfo(TimeTrackable, SavingUser, SoftDeletable):
     u_level = models.CharField(max_length=10, null=True, blank=True)
     u_height = models.CharField(max_length=10, null=True, blank=True)
     data_center = models.ForeignKey(DataCenter, null=True, blank=False)
-    server_room = models.ForeignKey(ServerRoom, null=True, blank=True)
-    rack = models.ForeignKey(Rack, null=True)
+    server_room = models.ForeignKey(ServerRoom, null=True, blank=False)
+    rack = models.ForeignKey(Rack, null=True, blank=True)
     # deperecated field, use rack instead
     rack_old = models.CharField(max_length=10, null=True, blank=True)
     slot_no = models.IntegerField(
@@ -953,6 +975,59 @@ class DeviceInfo(TimeTrackable, SavingUser, SoftDeletable):
         choices=Orientation(),
         default=Orientation.front.id,
     )
+
+    def clean(self):
+        """
+        Constraints:
+        - picked server-room is from picked data-center
+        - picked rack is from picked server-room
+        - postion = 0: orientation(left, right)
+        - postion > 0: orientation(front, middle, back)
+        - position <= rack.max_u_height
+        - slot_no: asset is_blade=True
+        """
+        Orientation.is_depth(1)
+        if self.server_room and self.data_center:
+            if self.server_room.data_center != self.data_center:
+                msg = '"{}" server room is not from "{}" datacenter'.format(
+                    self.server_room, self.data_center,
+                )
+                raise ValidationError(msg, code=INVALID_SERVER_ROOM)
+        if self.rack and self.server_room:
+            if self.rack.server_room != self.server_room:
+                msg = '"{}" rack is not from "{}" server room'.format(
+                    self.rack, self.server_room,
+                )
+                raise ValidationError(msg, code=INVALID_RACK)
+
+        if self.position == 0 and not Orientation.is_width(self.orientation):
+            msg = 'Valid orientations for picked position are: {}'.format(
+                ', '.join(
+                    getattr(Orientation, key).desc for key in ['left', 'right']
+                )
+            )
+            raise ValidationError(msg, code=INVALID_ORIENTATION)
+        if self.position > 0 and not Orientation.is_depth(self.orientation):
+            msg = 'Valid orientations for picked position are: {}'.format(
+                ', '.join(
+                    getattr(Orientation, key).desc
+                    for key in ['front', 'back', 'middle']
+                )
+            )
+            raise ValidationError(msg, code=INVALID_ORIENTATION)
+        if self.position > self.rack.max_u_height:
+            msg = 'Position is higher than "max u height" = {}'.format(
+                self.rack.max_u_height,
+            )
+            raise ValidationError(msg, code=INVALID_POSITION)
+        try:
+            is_blade = self.asset.model.category.is_blade
+        except AttributeError:
+            is_blade = False
+        else:
+            if is_blade and not self.slot_no:
+                msg = "'slot number' is required when asset is blade"
+                raise ValidationError(msg, code=REQUIRED_SLOT_NUMBER)
 
     @property
     def size(self):